requirements.txt
# Core data stack
pandas==2.2.3
numpy==1.26.4

# Visualization and dashboards
plotly==5.24.1
dash==2.18.2
dash-bootstrap-components==1.7.1

# Streamlit interface
streamlit==1.39.0

# Data handling
duckdb==1.1.1
pyarrow==17.0.0

# Automation & AI connectors
requests==2.32.3
openai==1.46.0

# Development tools
ipykernel==6.29.5
notebook==7.2.2
python-dotenv==1.0.1
kaleido==0.2.1